+++ EverlyticPush/EverlyticPush.Abstract/Delegates.cs
@@ -5,4 +5,6 @@ namespace Com.EverlyticPush.Abstract
     public delegate void OnResultReceivedDelegate(EvResult result);
 
     public delegate void OnNotificationHistoryResultsDelegate(List<EverlyticNotification> notifications);
+
+    public delegate void OnNotificationHistoryCountReceivedDelegate(int count);
 }