+++ EverlyticPush/EverlyticPush.Abstract/EverlyticNotification.cs
@@ -1,14 +1,16 @@
 using System;
+using System.Collections.Generic;
 
 namespace Com.EverlyticPush.Abstract
 {
     public class EverlyticNotification
     {
-        public string Body;
-        public DateTime DismissedAt;
-        public long MessageId;
-        public DateTime ReadAt;
-        public DateTime ReceivedAt;
         public string Title;
+        public string Body;
+        public DateTime? DismissedAt;
+        public long MessageId;
+        public DateTime? ReadAt;
+        public DateTime? ReceivedAt;
+        //public Dictionary<string, string> CustomAttributes;
     }
 }